introduce replacement Surrogate class, misnamed Facade extends it
